Release 1.1-0
On windows, there were still problems with python3 and the line ending
conversions. This is hopefully improved.

In the windows version, GetStdHandle did nog always return the same
value as the handle of e.g. stdout. Using the latter gave better results.

Added a function, pithon.set.default.executable, with which the default
python executable can be set for all instances that are started, unless
another value is set for the instance using pithon.set.executable.

On startup, the function 'find_python_cmd()' from the package 'findpython'
is used to try to determine a python executable that can be used.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rpithon@6 edb9625f-4e0d-4859-8d74-9fd3b1da38cb